FIXED MAGE-6549: Inventory: Setting "Set Items' Status to be In Stock When Order is Cancelled" does not work on website scope
FIXES #53

The problem is, that the `\Mage_CatalogInventory_Model_Stock_Item::getCanBackInStock()` is called inside the admin store context, when a order is canceled. Therefore, the value from the website is not used.

To use the correct node in the configuration we need to pass the storeId down to the `getStoreConfig()` call inside of `\Mage_CatalogInventory_Model_Stock_Item::getCanBackInStock()`. To not break the backwardscompatibility, all the parameters are optional.leStoreMode