#234: Remove int cast in isOrderIncrementIdUsed
Cast to int in the method isOrderIncrementIdUsed in resource model sales/quote is causing problems when using alphanumeric prefixes or increment models. Int cast is inconsistent too, since the column type this is checked against is varchar(50). The cast to int was replaced with a cast to string to solve this.